app/requirements.txt
PyQt5>=5.15
numpy>=2.0
matplotlib>=3.5
scipy>=1.10
Pillow>=10.0
